quicksearch/requirements.txt
googlesearch-python
customtkinter